crawlers/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
mysql-connector-python>=8.3.0
lxml>=5.1.0
